添加删除前的简单校验

+++ src/views/NewWorld/ArticleInsideReplyManage/ArticleInsideReplyManage.vue
@@ -320,6 +320,10 @@ export default {
                 return item["id"]
             });
 
+            if(!checkedData.length){
+                return this.$swal("请选择要操作的数据");
+            }
+
             this.$http.delete("insideReply", {
                     data : {
                         arr : checkedData
