+++ src/views/NewWorld/ArticleManage/ArticleManage.vue
@@ -201,6 +201,10 @@ export default {
                 return item["id"]
             });
 
+            if(!checkedData.length){
+                return this.$swal("请选择要操作的数据");
+            }
+
             this.$http.delete("article", {
                     data : {
                         arr : checkedData
