+++ src/views/NewWorld/ArticleTypeManage/ArticleTypeManage.vue
@@ -243,6 +243,10 @@ export default {
                 return item["id"]
             });
 
+            if(!checkedData.length){
+                return this.$swal("请选择要操作的数据");
+            }
+
             this.$http.delete("articleType", {
                     data : {
                         arr : checkedData
