+++ src/views/NewWorld/WallManage/WallManage.vue
@@ -293,6 +293,10 @@ export default {
                 return item["id"]
             });
 
+            if(!checkedData.length){
+                return this.$swal("请选择要操作的数据");
+            }
+
             this.$http.delete("wall", {
                     data : {
                         arr : checkedData
